tests/wpt/tests/css/css-text/parsing/word-break-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text: getComputedStyle().wordBreak</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-word-break">
<meta name="assert" content="word-break computed value is specified keyword.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("word-break", "normal");
test_computed_value("word-break", "keep-all");
test_computed_value("word-break", "break-all");
test_computed_value("word-break", "break-word");
test_computed_value("word-break", "auto-phrase");
</script>
</body>
</html>
